+++ src/aunmag/nightingale/utilities/TimerDone.java
@@ -2,11 +2,11 @@ package aunmag.nightingale.utilities;
 
 public class TimerDone {
 
-    private int timeDuration = 0;
-    private long timeInitial = 0;
-    private long timeTarget = 0;
+    private double timeDuration = 0.0;
+    private double timeInitial = 0.0;
+    private double timeTarget = 0.0;
 
-    public TimerDone(int timeDuration) {
+    public TimerDone(double timeDuration) {
         setTimeDuration(timeDuration);
     }
 
@@ -14,33 +14,33 @@ public class TimerDone {
         timeTarget = timeInitial + timeDuration;
     }
 
-    public boolean calculateIsDone(long timeCurrent) {
+    public boolean calculateIsDone(double timeCurrent) {
         return timeTarget <= timeCurrent;
     }
 
     /* Setters */
 
-    public void setTimeDuration(int timeDuration) {
+    public void setTimeDuration(double timeDuration) {
         this.timeDuration = timeDuration;
         updateTimeTarget();
     }
 
-    public void setTimeInitial(long timeInitial) {
+    public void setTimeInitial(double timeInitial) {
         this.timeInitial = timeInitial;
         updateTimeTarget();
     }
 
     /* Getters */
 
-    public int getTimeDuration() {
+    public double getTimeDuration() {
         return timeDuration;
     }
 
-    public long getTimeInitial() {
+    public double getTimeInitial() {
         return timeInitial;
     }
 
-    public long getTimeTarget() {
+    public double getTimeTarget() {
         return timeTarget;
     }
